Maintenance: refactored auth into its own module
